Совместные даты - сделано  вычисление совместного количества прожитых дней  выбор персоны в активности-диалоге с сортировкой  расчёт даты по количеству совместно прожитых дней  поворот экрана расчёта совместных дат  отлажено для Андроид 4.0.3  Сделан класс Person, чтобы в запросах сразу выдавать все поля  Начал FindDateActivity делать для  поиска совместных дат  Пока через SimpleAdapter
(cherry picked from commit 8f5c6313625397f6a669c9425f5aacf114b6bda3)

+++ jubelee_dialog/src/main/java/ru/bartex/jubelee_dialog/Person.java
@@ -64,7 +64,7 @@ public class Person implements Serializable {
         return person_id;
     }
 
-    public void setPerson_id(int person_id) {
+    public void setPerson_id(long person_id) {
         this.person_id = person_id;
     }
